Replace RTM Bot with SocketModeAdapter
RTM Client is somewhat depreciated, and to continue to use Slack bots we
want to have a new client that mimics what the RTMClient gave us but
supporting the new socket mode.

We create a new SocketAdapter that provides us a lot of the API previously
provided by the RTMClient and we simplify our BotClient to use that new
adapter.

We change our command dispatch to filter on a new 'command' queue so
that we can reuse more of our dispatch code instead of having a similar
duplicate like our old RTMClient.

Since a few values change (like data -> messages) we also update some of
our default plugins/decorators to work with the new structure.